fix: rounding error when applying fee
When we compute the final volumeEth, we do a division, but if this
division ends being inexact, the default behaviour was to truncate the
number. This is problematic since we end up sending less eth that
required to perform the trade.

Now, when do a divCeil or div depending what we need.