feat: ask allowance only once for max value
In order to save gas, we ask for allowance only the first time we do a
sell with a given Token; instead of doing everytime.

When we ask for allowance we ask for the maximum possible value:
2^256 -1. And then, before every trade we check we still have allowance
left, if not, we ask again.

SellSteps have changed, we only show the 'allowance steps' when we do
ask for it. If not, we omit them.

Also, since we ask allowance for the maxium, insted the requested
volume, we don't show the volume anymore in the sell step.

resolves #52